refactor: trim duplicated router and wallet helpers
